Implemented on-the-fly dependency collection

+++ src/bin/cargo-rustyrts.rs
@@ -1,23 +1,21 @@
 use itertools::Itertools;
-use rustyrts::checksums::Checksums;
 use rustyrts::constants::{
-    DESC_FLAG, ENDING_CHANGES, ENDING_CHECKSUM, ENDING_GRAPH, ENDING_PROCESS_TRACE, ENDING_TEST,
+    DESC_FLAG, ENDING_CHANGES, ENDING_PROCESS_TRACE, ENDING_TEST,
     ENDING_TRACE, ENV_RUSTC_WRAPPER, ENV_RUSTYRTS_ARGS, ENV_RUSTYRTS_MODE, ENV_RUSTYRTS_VERBOSE,
-    ENV_SKIP_ANALYSIS, ENV_TARGET_DIR, FILE_COMPLETE_GRAPH, VERBOSE_COUNT,
+    ENV_SKIP_ANALYSIS, ENV_TARGET_DIR,VERBOSE_COUNT,
 };
 use rustyrts::fs_utils::{
-    get_dynamic_path, get_static_path, get_target_dir, read_lines, read_lines_filter_map,
+    get_dynamic_path, get_static_path, get_target_dir, read_lines, 
 };
-use rustyrts::static_rts::graph::DependencyGraph;
 use rustyrts::utils;
+use rustyrts::constants::ENDING_DEPENDENCIES;
 use serde_json;
 use std::collections::HashSet;
 use std::ffi::OsString;
 use std::fs::{
-    create_dir_all, read, read_dir, read_to_string, remove_dir_all, remove_file, DirEntry,
-    OpenOptions,
+    create_dir_all,  read_dir, read_to_string, remove_dir_all, remove_file, DirEntry,
+    
 };
-use std::io::Write;
 use std::path::PathBuf;
 use std::process::Command;
 use std::str::FromStr;
@@ -111,7 +109,7 @@ fn cargo() -> Command {
     Command::new(std::env::var_os("CARGO").unwrap_or_else(|| OsString::from("cargo")))
 }
 
-#[derive(PartialEq)]
+#[derive(PartialEq, Clone, Copy)]
 enum Mode {
     Clean,
     Dynamic,
@@ -320,19 +318,13 @@ fn main() {
             Some(Mode::Clean) => {
                 clean();
             }
-            Some(Mode::Static) => {
-                // This arm is for when `cargo rustyrts static` is called. We call `cargo build`,
+            Some(mode) => {
+                // This arm is for when `cargo rustyrts (static|dynamic)` is called. We call `cargo build`,
                 // but with the `RUSTC` env var set to the `cargo-rustyrts` binary so that we come back in the other branch,
-                // and dispatch the invocations to `rustyrts-static`, respectively.
-                run_cargo_rustc_static();
-                select_and_execute_tests_static();
-            }
-            Some(Mode::Dynamic) => {
-                // This arm is for when `cargo rustyrts dynamic` is called. We call `cargo build`,
-                // but with the `RUSTC` env var set to the `cargo-rustyrts` binary so that we come back in the other branch,
-                // and dispatch the invocations to `rustyrts-dynamic`, respectively.
-                run_cargo_rustc_dynamic();
-                select_and_execute_tests_dynamic();
+                // and dispatch the invocations to `rustyrts-(static|dnyamic)`, respectively.
+                run_cargo_build(mode);
+                select_and_execute_tests(mode);
+                
             }
             _ => {
                 show_error(
@@ -418,201 +410,17 @@ fn run_rustyrts() {
 }
 
 //######################################################################################################################
-// STATIC RTS
+// GENERIC RTS
 
 /// This will construct and execute a command like:
-/// `cargo build --bin some_crate_name -v -- --top_crate_name some_top_crate_name --domain interval -v cargo-rustyrts-marker-end`
-/// using the rustc wrapper for static rustyrts
-fn run_cargo_rustc_static() {
-    let path_buf: PathBuf = get_static_path(false);
-
-    create_dir_all(path_buf.as_path()).expect(&format!(
-        "Failed to create directory {}",
-        path_buf.display()
-    ));
-
-    let files = read_dir(path_buf.as_path()).unwrap();
-    for path_res in files {
-        if let Ok(path) = path_res {
-            if path.file_name().to_str().unwrap().ends_with(ENDING_CHANGES) {
-                remove_file(path.path()).unwrap();
-            }
-        }
-    }
-
-    let cmd = cargo_build(Mode::Static);
-    execute(cmd);
-}
-
-/// This will construct and execute a command like:
-/// * `cargo test --no-fail-fast -- --exact test_1 test_2 ...` (If some tests are affected)
-/// * `cargo test --no-fail-fast --no-run` (If no tests are affected)
-/// using the rustc wrapper for static rustyrts
-fn select_and_execute_tests_static() {
-    let verbose = has_arg_flag("-v");
-
-    let path_buf = get_static_path(true);
-
-    let files: Vec<DirEntry> = read_dir(path_buf.as_path())
-        .unwrap()
-        .map(|maybe_path| maybe_path.unwrap())
-        .collect();
-
-    // Read graphs
-    let mut dependency_graph: DependencyGraph<String> = DependencyGraph::new();
-    let edges = read_lines_filter_map(
-        &files,
-        ENDING_GRAPH,
-        |line| !line.trim_start().starts_with("\\") && line.contains("\" -> \""),
-        |line| line,
-    );
-    dependency_graph.import_edges(edges);
-
-    if has_arg_flag("-vv") {
-        let mut complete_graph_path = path_buf.clone();
-        complete_graph_path.push(FILE_COMPLETE_GRAPH);
-        let mut file = match OpenOptions::new()
-            .create(true)
-            .write(true)
-            .append(false)
-            .open(complete_graph_path.as_path())
-        {
-            Ok(file) => file,
-            Err(reason) => panic!("Failed to open file: {}", reason),
-        };
-
-        let checksum_files = files.iter().filter(|path| {
-            path.file_name()
-                .to_str()
-                .unwrap()
-                .ends_with(ENDING_CHECKSUM)
-        });
-        let mut checksums_nodes = HashSet::new();
-
-        for checkums_path in checksum_files {
-            let maybe_checksums = read(checkums_path.path());
-            if let Ok(checksums) = maybe_checksums {
-                let checksums = Checksums::from(checksums.as_slice());
-                for node in checksums.keys() {
-                    checksums_nodes.insert(node.clone());
-                }
-            }
-        }
-
-        match file.write_all(format!("{}\n", dependency_graph.pretty(checksums_nodes)).as_bytes()) {
-            Ok(_) => {}
-            Err(reason) => panic!("Failed to write to file: {}", reason),
-        };
-    }
-
-    // Read changed nodes
-    let changed_nodes = read_lines_filter_map(
-        &files,
-        ENDING_CHANGES,
-        |line| line != "" && dependency_graph.get_node(&line).is_some(),
-        |line| dependency_graph.get_node(&line).unwrap(),
-    );
-
-    if verbose {
-        println!(
-            "Nodes that have changed:\n{}\n",
-            changed_nodes.iter().sorted().join(", ")
-        );
-    } else {
-        println!(
-            "#Nodes that have changed: {}\n",
-            changed_nodes.iter().count()
-        );
-    }
-
-    // Read possibly affected tests
-    let tests = read_lines_filter_map(
-        &files,
-        "test",
-        |line| line != "" && dependency_graph.get_node(&line).is_some(),
-        |line| dependency_graph.get_node(&line).unwrap(),
-    );
-
-    println!("#Tests that have been found: {}\n", tests.iter().count());
-
-    #[cfg(not(feature = "print_paths"))]
-    {
-        let reached_nodes = dependency_graph.reachable_nodes(changed_nodes);
-        let affected_tests: HashSet<&&String> = tests.intersection(&reached_nodes).collect();
-
-        println!(
-            "#Nodes that reach any changed node in the graph: {}\n",
-            reached_nodes.iter().count()
-        );
-
-        if verbose {
-            println!(
-                "Affected tests:\n{}\n",
-                affected_tests.iter().sorted().join(", ")
-            );
-        } else {
-            println!("#Affected tests: {}\n", affected_tests.iter().count());
-        }
-
-        let cmd = cargo_test(Mode::Static, affected_tests.into_iter().map(|test| *test));
-
-        execute(cmd);
-    }
-    #[cfg(feature = "print_paths")]
-    {
-        let (reached_nodes, affected_tests) = dependency_graph.affected_tests(changed_nodes, tests);
-
-        println!(
-            "#Nodes that reach any changed node in the graph: {}\n",
-            reached_nodes.iter().count()
-        );
-
-        if verbose {
-            println!(
-                "Affected tests:\n{}\n",
-                affected_tests
-                    .iter()
-                    .sorted()
-                    .map(|(k, v)| {
-                        let path = format!(
-                            "{}: < {}{} >",
-                            k,
-                            v[0],
-                            if v.len() >= VERBOSE_COUNT {
-                                format!(
-                                    "<- ... <- {}",
-                                    v[v.len() - VERBOSE_COUNT..].iter().join(" <- ")
-                                )
-                            } else {
-                                v[1..].iter().join(" <- ")
-                            }
-                        );
-
-                        format!("{}: {}", k, path)
-                    })
-                    .join("\n")
-            );
-        } else {
-            println!("#Affected tests: {}\n", affected_tests.iter().count());
-        }
-
-        let cmd = cargo_test(
-            Mode::Static,
-            affected_tests.keys().into_iter().map(|test| *test),
-        );
-
-        execute(cmd);
-    }
-}
-
-//######################################################################################################################
-// DYNAMIC RTS
-
-/// This will construct and execute a command like:
-/// `cargo build --bin some_crate_name -v -- cargo-rustyrts-marker-begin --top_crate_name some_top_crate_name --domain interval -v cargo-rustyrts-marker-end`
+/// `cargo build --bin some_crate_name -v -- cargo-rustyrts-marker-begin --top_crate_name some_top_crate_name --domain interval -v`
 /// using the rustc wrapper for dynamic rustyrts
-fn run_cargo_rustc_dynamic() {
-    let path_buf: PathBuf = get_dynamic_path(false, None);
+fn run_cargo_build(mode: Mode){
+    let path_buf = match mode {
+        Mode::Clean => unreachable!(),
+        Mode::Dynamic => get_dynamic_path(false, None),
+        Mode::Static => get_static_path(false),
+    };
 
     create_dir_all(path_buf.as_path()).expect(&format!(
         "Failed to create directory {}",
@@ -628,8 +436,10 @@ fn run_cargo_rustc_dynamic() {
             }
 
             #[cfg(unix)]
-            if file_name.to_str().unwrap().ends_with(ENDING_PROCESS_TRACE) {
-                remove_file(path.path()).unwrap();
+            if mode == Mode::Dynamic {
+                if file_name.to_str().unwrap().ends_with(ENDING_PROCESS_TRACE) {
+                    remove_file(path.path()).unwrap();
+                }
             }
         }
     }
@@ -637,18 +447,22 @@ fn run_cargo_rustc_dynamic() {
     // Now we run `cargo build $FLAGS $ARGS`, giving the user the
     // chance to add additional arguments. `FLAGS` is set to identify
     // this target.  The user gets to control what gets actually passed to rustyrts.
-    let cmd = cargo_build(Mode::Dynamic);
+    let cmd = cargo_build(mode);
     execute(cmd);
 }
 
 // This will construct command line like:
 // either   `cargo test --no-fail-fast -- --exact test_1 test_2 ...` (If some tests are affected)
 // or       `cargo test --no-fail-fast --no-run` (If no tests are affected)
-/// using the rustc wrapper for dynamic rustyrts
-fn select_and_execute_tests_dynamic() {
+/// using the respective rustc wrapper
+fn select_and_execute_tests(mode: Mode) {
     let verbose = has_arg_flag("-v");
 
-    let path_buf = get_dynamic_path(true, None);
+    let path_buf = match mode {
+        Mode::Clean => unreachable!(),
+        Mode::Dynamic => get_dynamic_path(true, None),
+        Mode::Static => get_static_path(true),
+    };
 
     let files: Vec<DirEntry> = read_dir(path_buf.as_path())
         .unwrap()
@@ -675,9 +489,16 @@ fn select_and_execute_tests_dynamic() {
 
     println!("#Tests that have been found: {}\n", tests.iter().count());
 
-    // Read traces
+    // Read traces or dependencies
+    let ending = match mode {
+        Mode::Clean => unreachable!(),
+        Mode::Dynamic => ENDING_TRACE,
+        Mode::Static => ENDING_DEPENDENCIES,
+    };
+
     let mut affected_tests: Vec<(String, Option<HashSet<String>>)> = Vec::new();
-    let traced_tests: Vec<&DirEntry> = files
+
+    let analyzed_tests: Vec<&DirEntry> = files
         .iter()
         .filter(|traces| {
             traces
@@ -685,11 +506,11 @@ fn select_and_execute_tests_dynamic() {
                 .to_os_string()
                 .into_string()
                 .unwrap()
-                .ends_with(ENDING_TRACE)
+                .ends_with(ending)
         })
         .collect();
 
-    let traced_tests_names: HashSet<String> = traced_tests
+    let analyzed_tests_names: HashSet<String> = analyzed_tests
         .iter()
         .map(|f| {
             f.file_name()
@@ -707,18 +528,18 @@ fn select_and_execute_tests_dynamic() {
         .collect();
 
     println!(
-        "#Tests with traces:: {}\n",
-        traced_tests_names.iter().count()
+        "#Tests with information: {}\n",
+        analyzed_tests_names.iter().count()
     );
 
     affected_tests.append(
         &mut tests
-            .difference(&traced_tests_names)
+            .difference(&analyzed_tests_names)
             .map(|s| (s.clone(), None))
             .collect_vec(),
     );
 
-    for file in traced_tests {
+    for file in analyzed_tests {
         let traced_nodes: HashSet<String> = read_to_string(file.path())
             .unwrap()
             .split("\n")
@@ -775,6 +596,6 @@ fn select_and_execute_tests_dynamic() {
         println!("#Affected tests: {}\n", affected_tests.iter().count());
     }
 
-    let cmd = cargo_test(Mode::Dynamic, affected_tests.iter().map(|(n, _)| n));
+    let cmd = cargo_test(mode, affected_tests.iter().map(|(n, _)| n));
     execute(cmd);
 }
